feat: delegate restCountriesAPI to its own files
