+++ app/src/main/java/com/example/valorant_app/data/hiltmodules/ValorantNetworkModule.kt
@@ -1,13 +1,12 @@
 package com.example.valorant_app.data.hiltmodules
 
-import com.example.valorant_app.data.services.ApiService
+import com.example.valorant_app.data.services.ValorantApiService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
 
 
 @Module
@@ -23,28 +22,7 @@ object ValorantNetworkModule {
     }
 
     @Provides
-    fun provideApiService(retrofit: Retrofit): ApiService {
-        return retrofit.create(ApiService::class.java)
-    }
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-object RestCountriesModule {
-
-    @Provides
-    @Singleton
-    fun provideRestCountriesRetrofit(): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl("https://restcountries.com/")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-
-    @Provides
-    @Singleton
-    fun provideRestCountriesApi(retrofit: Retrofit): ApiService {
-        return retrofit.create(ApiService::class.java)
+    fun provideApiService(retrofit: Retrofit): ValorantApiService {
+        return retrofit.create(ValorantApiService::class.java)
     }
 }
