+++ app/src/main/java/com/example/valorant_app/data/repository/WeaponRepository.kt
@@ -2,7 +2,7 @@ package com.example.valorant_app.data.repository
 
 import com.example.valorant_app.data.entities.WeaponCard
 import com.example.valorant_app.data.services.ApiResponse
-import com.example.valorant_app.data.services.ApiService
+import com.example.valorant_app.data.services.ValorantApiService
 import javax.inject.Inject
 
 interface WeaponRepository {
@@ -10,11 +10,11 @@ interface WeaponRepository {
 }
 
 class WeaponRepositoryImpl @Inject constructor(
-    private val apiService: ApiService
+    private val valorantApiService: ValorantApiService
 ) : WeaponRepository {
     override suspend fun getAllWeaponCard(): ApiResponse<List<WeaponCard>> {
         return try {
-            val response = apiService.getAllWeaponCard()
+            val response = valorantApiService.getAllWeaponCard()
             ApiResponse(
                 status = response.status,
                 data = response.data
